requirements.txt
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
Django==4.3.11
django-bootstrap-v5==1.0.11